requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.5
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-email-verification==0.0.9.post1
django-environ==0.7.0
django-filter==2.4.0
django-jazzmin==2.4.8
django-money==2.1
django-summernote==0.8.11.6
django-timezone-field==4.2.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.6.4
incremental==21.3.0
kombu==5.1.0
msgpack==1.0.2
Pillow==8.3.1
prompt-toolkit==3.0.20
py-moneyed==1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.0
zope.interface==5.4.0
